Allow exporting files from HTTP URLs by downloading them
