+++ cmd/forklift/dev/plt/downloads.go
@@ -1,0 +1,36 @@
+package plt
+
+import (
+	"fmt"
+
+	"github.com/urfave/cli/v2"
+
+	fcli "github.com/PlanktoScope/forklift/internal/app/forklift/cli"
+)
+
+// cache-dl
+
+func cacheDlAction(versions Versions) cli.ActionFunc {
+	return func(c *cli.Context) error {
+		pallet, repoCache, dlCache, err := processFullBaseArgs(c, true, true)
+		if err != nil {
+			return err
+		}
+		if err = fcli.CheckCompatibility(
+			pallet, repoCache, versions.Tool, versions.MinSupportedRepo, versions.MinSupportedPallet,
+			c.Bool("ignore-tool-version"),
+		); err != nil {
+			return err
+		}
+
+		fmt.Println("Downloading files to be exported by the development pallet...")
+		if err := fcli.DownloadExportFiles(
+			0, pallet, repoCache, dlCache, false, c.Bool("parallel"),
+		); err != nil {
+			return err
+		}
+		fmt.Println()
+		fmt.Println("Done!")
+		return nil
+	}
+}
