+++ cmd/forklift/dev/plt/repositories.go
@@ -12,7 +12,7 @@ import (
 
 func cacheRepoAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		pallet, cache, err := processFullBaseArgs(c, false, false)
+		pallet, cache, _, err := processFullBaseArgs(c, false, false)
 		if err != nil {
 			return err
 		}
@@ -53,7 +53,7 @@ func lsRepoAction(c *cli.Context) error {
 // show-repo
 
 func showRepoAction(c *cli.Context) error {
-	pallet, cache, err := processFullBaseArgs(c, true, true)
+	pallet, cache, _, err := processFullBaseArgs(c, true, true)
 	if err != nil {
 		return err
 	}
@@ -65,7 +65,7 @@ func showRepoAction(c *cli.Context) error {
 
 func addRepoAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		pallet, repoCache, err := processFullBaseArgs(c, false, false)
+		pallet, repoCache, dlCache, err := processFullBaseArgs(c, false, false)
 		if err != nil {
 			return err
 		}
@@ -83,7 +83,9 @@ func addRepoAction(versions Versions) cli.ActionFunc {
 		}
 
 		if !c.Bool("no-cache-req") {
-			if _, err = fcli.CacheStagingRequirements(pallet, repoCache.Path()); err != nil {
+			if err = fcli.CacheStagingRequirements(
+				pallet, repoCache.Path(), repoCache, dlCache, false, c.Bool("parallel"),
+			); err != nil {
 				return err
 			}
 		}
