+++ cmd/forklift/plt/deployments.go
@@ -9,7 +9,7 @@ import (
 // ls-depl
 
 func lsDeplAction(c *cli.Context) error {
-	pallet, cache, err := processFullBaseArgs(c.String("workspace"), true)
+	pallet, cache, _, err := processFullBaseArgs(c.String("workspace"), true)
 	if err != nil {
 		return err
 	}
@@ -20,7 +20,7 @@ func lsDeplAction(c *cli.Context) error {
 // show-depl
 
 func showDeplAction(c *cli.Context) error {
-	pallet, cache, err := processFullBaseArgs(c.String("workspace"), true)
+	pallet, cache, _, err := processFullBaseArgs(c.String("workspace"), true)
 	if err != nil {
 		return err
 	}
@@ -32,7 +32,7 @@ func showDeplAction(c *cli.Context) error {
 // locate-depl-pkg
 
 func locateDeplPkgAction(c *cli.Context) error {
-	pallet, cache, err := processFullBaseArgs(c.String("workspace"), true)
+	pallet, cache, _, err := processFullBaseArgs(c.String("workspace"), true)
 	if err != nil {
 		return err
 	}
