+++ cmd/forklift/plt/images.go
@@ -12,7 +12,7 @@ import (
 
 func cacheImgAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		pallet, cache, err := processFullBaseArgs(c.String("workspace"), true)
+		pallet, cache, _, err := processFullBaseArgs(c.String("workspace"), true)
 		if err != nil {
 			return err
 		}
