+++ cmd/forklift/plt/repositories.go
@@ -12,7 +12,7 @@ import (
 
 func cacheRepoAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		pallet, cache, err := processFullBaseArgs(c.String("workspace"), false)
+		pallet, cache, _, err := processFullBaseArgs(c.String("workspace"), false)
 		if err != nil {
 			return err
 		}
@@ -54,7 +54,7 @@ func lsRepoAction(c *cli.Context) error {
 // show-repo
 
 func showRepoAction(c *cli.Context) error {
-	pallet, cache, err := processFullBaseArgs(c.String("workspace"), true)
+	pallet, cache, _, err := processFullBaseArgs(c.String("workspace"), true)
 	if err != nil {
 		return err
 	}
@@ -66,7 +66,7 @@ func showRepoAction(c *cli.Context) error {
 
 func addRepoAction(versions Versions) cli.ActionFunc {
 	return func(c *cli.Context) error {
-		pallet, repoCache, err := processFullBaseArgs(c.String("workspace"), false)
+		pallet, repoCache, dlCache, err := processFullBaseArgs(c.String("workspace"), false)
 		if err != nil {
 			return err
 		}
@@ -82,7 +82,9 @@ func addRepoAction(versions Versions) cli.ActionFunc {
 		}
 
 		if !c.Bool("no-cache-req") {
-			if _, err = fcli.CacheStagingRequirements(pallet, repoCache.Path()); err != nil {
+			if err = fcli.CacheStagingRequirements(
+				pallet, repoCache.Path(), repoCache, dlCache, false, c.Bool("parallel"),
+			); err != nil {
 				return err
 			}
 		}
